tools/calendar/requirements.txt
caldav>=2.2.0
icalendar>=6.0.0
pyyaml>=6.0
